pyproject.toml
[project]
name = "personal-chef"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosmtplib>=3.0.1",
    "email-validator>=2.0.0",
    "fastapi>=0.133.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "langgraph>=0.2.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.36",
    "streamlit>=1.54.0",
    "tavily-python>=0.7.22",
    "uvicorn[standard]>=0.41.0",
]
